fractional-worker/requirements.txt
temporalio>=1.7.0
httpx>=0.25.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
